+++ src/QueryType/ManagedResources/Query/Command/Stopwords/Add.php
@@ -7,16 +7,14 @@
  * file that was distributed with this source code.
  */
 
-namespace Solarium\QueryType\ManagedResources\Query\Stopwords\Command;
+namespace Solarium\QueryType\ManagedResources\Query\Command\Stopwords;
 
-use Solarium\Core\Client\Request;
-use Solarium\QueryType\ManagedResources\Query\AbstractCommand;
-use Solarium\QueryType\ManagedResources\Query\Stopwords;
+use Solarium\QueryType\ManagedResources\Query\Command\AbstractAdd;
 
 /**
  * Add.
  */
-class Add extends AbstractCommand
+class Add extends AbstractAdd
 {
     /**
      * Stopwords to add.
@@ -24,26 +22,6 @@ class Add extends AbstractCommand
      * @var array
      */
     protected $stopwords;
-
-    /**
-     * Returns command type, for use in adapters.
-     *
-     * @return string
-     */
-    public function getType(): string
-    {
-        return Stopwords::COMMAND_ADD;
-    }
-
-    /**
-     * Returns request method.
-     *
-     * @return string
-     */
-    public function getRequestMethod(): string
-    {
-        return Request::METHOD_PUT;
-    }
 
     /**
      * Get stopwords.
@@ -70,22 +48,12 @@ class Add extends AbstractCommand
     }
 
     /**
-     * Returns the data to be sent to Solr.
+     * Returns the raw data to be sent to Solr.
      *
      * @return string
      */
     public function getRawData(): string
     {
         return json_encode($this->stopwords);
-    }
-
-    /**
-     * Returns the term to be sent to Solr.
-     *
-     * @return string
-     */
-    public function getTerm(): string
-    {
-        return '';
     }
 }