+++ src/QueryType/ManagedResources/Query/Resources.php
@@ -10,7 +10,7 @@
 namespace Solarium\QueryType\ManagedResources\Query;
 
 use Solarium\Core\Client\Client;
-use Solarium\Core\Query\AbstractQuery;
+use Solarium\Core\Query\AbstractQuery as BaseQuery;
 use Solarium\Core\Query\RequestBuilderInterface;
 use Solarium\Core\Query\ResponseParserInterface;
 use Solarium\QueryType\ManagedResources\RequestBuilder\Resources as RequestBuilder;
@@ -20,7 +20,7 @@ use Solarium\QueryType\ManagedResources\Result\Resources\ResourceList;
 /**
  * Resources.
  */
-class Resources extends AbstractQuery
+class Resources extends BaseQuery
 {
     /**
      * Default options.
@@ -63,7 +63,7 @@ class Resources extends AbstractQuery
     /**
      * Get the request builder class for this query.
      *
-     * @return RequestBuilder
+     * @return \Solarium\QueryType\ManagedResources\RequestBuilder\Resources
      */
     public function getRequestBuilder(): RequestBuilderInterface
     {
@@ -73,7 +73,7 @@ class Resources extends AbstractQuery
     /**
      * Get the response parser class for this query.
      *
-     * @return ResponseParser
+     * @return \Solarium\QueryType\ManagedResources\ResponseParser\Resources
      */
     public function getResponseParser(): ResponseParserInterface
     {