+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 ### Changed
+- Refactored Managed Resources code: use `createCommand()` and `createInitArgs()` to issue commands
 
 ### Removed